Simplify arguments setup for embedding script
Moves common arguments across all subparsers to the top-level parser and
reorganizes the corresponding commands in the flu Snakefile to use this new
format for calling the embed script.

This commit also adds a `required=True` keyword argument to the subparsers
constructor, so we do not need to manually check for the existence of the
command argument. Also, removes checks for specific supported subcommand names
since argparse already does this for us.