Generalize Snakemake rules for seasonal flu
Replace per-method rules for scatterplots, etc. with a single rule for each type
of analysis and encode method information with wildcards. As part of this
change, this commit also introduces an `--embedding-columns` argument for the
within/between group script such that the coordinate column names for embeddings
are explicitly defined by the user instead of guessed at by the program.