Fix bug with scaled distance in within/between plots
We only need to apply the standard scaler to the precalculated Euclidean
distance field, so we can plot the distances on the same scale as the
SVM threshold. The original code was recalculating Euclidean distances
for all numerical fields in a given embedding data frame (including the
HDBSCAN label!) and then scaling that distance. This led to subtle bugs
in SVM threshold accuracy for t-SNE and UMAP, but obvious issues with
PCA (an embedding with 10 PC fields and a cluster label field).

This commit also changes the classifier threshold output to a scalar
from an array for readability.