Add support for interactive and static figures
Uses the same Markdown figure caption and label to reference both a
static image in a PDF and an interactive image in HTML. This hack works
because a) the PDF document ignores the iframe including the interactive
image and only displays the static image and b) the HTML document
selectively hides all images with the CSS class "static-embedding"
without hiding the corresponding caption. This latter step only works
because we use the "link_attributes" pandoc extension during compilation
and add a ".static-embedding" to the static image label.

This commit adds an example for this type of dual embedded figure for
just the flu results. Similar changes to Zika and MERS image labels need
to happen and the corresponding PNGs needed to be added.