Update HA+NA notebook to display all clades
Notebook was hiding clades that did not have colors assigned (no entry
in the "domain" or "range_" variables). This commit adds Nextstrain
colors to all clades in the analysis.

This commit also adds the updates auspice JSON and final HA+NA figure
for the updated Snakemake workflow where sequences are parsed, filtered,
deduplicated, and then paired.