Fix altair saver export issue with Jupyter notebook
The export of PNGs from the final Jupyter notebook in the seasonal flu
analysis did not work initially. This commit moves the altair_saver
package installation from pip to conda with the main altair and
chromedriver packages. This move fixes the export issue for me.

I also remove the `anaconda` channel since it is redundant with the
`default` channel (or should be).