Manually assign expanded multihost clades
Adds new manual clade annotations based on the automated annotations in
the workflow such that multiple hosts can be included in the same clade.
Also defines custom colors for these manual clades and the original
clades to overcome the maximum number of color values Auspice allows in
categorical fields and cleans up the Auspice config file to eliminate
all warnings.