Parameterize the flu workflow's random seed
Use a single variable to define the same random seed throughout the flu workflow
and also use this variable during the augur filter step. When we upgrade to
augur 13.0.1, the `--subsample-seed` argument will guarantee that we always
select the same strains for each run of the workflow. This change will make the
entire workflow more reproducible.