Initialize t-SNE with PCA
Follow standard practice for t-SNE embedding initialization by first
performing PCA on the given alignment and then using this PCA embedding
as a starting point for the t-SNE optimization. This behavior will be
the default in scikit-learn soon, but since we use a distance matrix as
our standard input to t-SNE embeddings, we need to precalculate the PCA
embedding ourselves.

This commit updates the logic of the embed script to use PCA for both
PCA and t-SNE embeddings. It also adds a `--components` argument to both
t-SNE and UMAP embeddings, for consistency with the other embeddings. We
use this components argument for both PCA and t-SNE when running the
t-SNE embedding. Instead of using the current scikit-learn default of a
"random" init, we now pass the PCA components for the requested number
of dimensions as the `init` argument. This change also requires the
Snakemake rule for t-SNE to accept an alignment as input.