Start work on unified workflow
Creates a top-level Snakefile that imports Snakefiles for each pathogen
analysis into a single unified workflow. This work starts with seasonal
flu, renaming paths for inputs/outputs, logs, scripts, etc. to work from
the top-level of the cartography directory. More work will be required
to rename rules to avoid conflicts when other pathogens are included.

This approach of a unified workflow provides several advantages over the
previous approach. Specifically, it allows us to:

 * automate passing the parameters from grid searches into the final
   embedding runs

 * re-generate figures and tables for the manuscript as needed with a
   single command

 * use a single Conda environment for all workflows instead of one
   environment per workflow