Add notebook to analyze MERS recombination events
Adds a Jupyter notebook to parse the extended NEXUS file from Muller et
al. and identify the recombination events encoded in this data
structure.

Note that we also update the "version" of the node data JSON file, so it
works with the latest version of Augur used by this project.