Do not use imputed alignment for any embeddings
While we originally wanted to play with the idea of using an imputed
alignment to see how much better the PCA embeddings would be, I don't
see a good justification for actually keeping this approach in practice.
To make SARS-CoV-2 analyses consistent with the MERS analyses, I've
dropped the rule to impute missing bases for the SARS-CoV-2 workflow.