Tune recombination rates
Tunes recombination rates for high, moderate, and low levels of
recombination to match expected value for MERS, seasonal CoV, and an
arbitrarily lower-but-nonzero recombination rate than seasonal CoV,
respectively. Uses a custom extension of SANTA-SIM to export whether a
sampled virus is a recombinant or not and samples every generation to
increase the chances of finding recombinants.