Define and filter a burn-in period for all simulations
Following the approach I took in the long-term flu forecasting
simulations, this commit updates the simulation configs for
coronavirus-like populations to output 10 + 5 years worth
generations (at 30 generations per year) and updates the workflow to
filter the earliest 10 years of simulated data as a "burn-in" period. In
this way, the flu and coronavirus populations produce 15 years of
simulated data and only the last 5 are used for the analysis.