Add and correct doctests for indels
Adds a simple doctest for two genomes with the same indel and a single
mismatch that should have a distance of 1 but currently has a distance
of 2 because the identical indel gets counted as a mismatch. Corrects
two other tests where the indel distance is not what we would expect.