Update indel distance logic
Updates logic for indel distance calculations to consider indels between
genomes as mismatched only when the indel start positions differ or the
same indel start position has an indel of a different length in the two
genomes. As part of this update, adds a function `find_ranges` that
returns a dictionary of gap lengths indexed by gap start positions and
rewrites `get_hamming_distances` to use the values for indel distance
calculations. These changes cause the doctests for the
`get_hamming_distances` function to pass.