Track distinct and total clusters per mutation
Reorganizes logic for finding cluster-specific mutations such that the
output table includes the mutation (position and derived allele), the
number of clusters that the mutation appears in, and the list of
distinct clusters with the mutation. Mutations that appear in all
clusters are omitted from output. Users can filter the output by number
of clusters equal to 1 to find mutations that only appear in a single
cluster and use the lists of distinct clusters to inspect shared
mutations between clusters.