Update influenza-like grid search results
Updates parameters from grid search based on new cross-validation with
MSE approach. Optimal parameters remain the same for UMAP, but they
change for the other methods. For PCA and MDS, CV selects substantially
fewer components than before. For t-SNE, CV selects a much higher
perplexity value. We may need to test a higher perplexity value, to
confirm that the value we've selected is an optimum.