Use fixed size train and test data for cross-validation
Instead of incrementally increasing the training data for each split in
the time series cross-validation, use the same size for both training
and test data in each split. This approach avoids adding bias and
variance to the accuracy estimates for each embedding method that can
come from changing the size of the input data. For example, t-SNE is
notoriously sensitive to number of input data points and initial CV
results exhibited the bias and increased variance for different t-SNE
parameters as the training sizes grew for each split.