Do not catch all exceptions in embed
Removes catch for all exceptions in the embed script. This kind of catch
obscures the true cause of errors and always requires me to go back to
the code and manually comment out the try/except logic. While this
software is still relatively young and officially unpublished, we should
allow exceptions to occur and then resolve the bugs that cause them in
the first place.