Try train/test with separate embeddings
Create train/test splits with separate test embeddings as the "observed"
data instead of the full embeddings. This approach tries to get around
the issue that the previous method of projecting the training embedding
into the full embedding with Procrustes analysis didn't work when the
embedding sizes were so different.