Use intermediate files for clade, date, and y axis
Instead of using Auspice JSONs that get created at the end of the
workflow to create a table of metadata that we used during clustering,
build that table from intermediate files with the same information. This
change removes circular logic where the Auspice JSON depended on the
final clustered embedding JSONs which in turn depended on the Auspice
JSON for metadata.