Pin Conda environment to known working versions
Pins Conda dependencies to known working versions. In the process, we
upgrade to Altair 5 and replace altair_saver with vl-convert (the new
recommended and default backend for Altair). Updates the early flu
notebook as an example of how saving figures changes with this latest
version of Altair and updates the corresponding flu figures.