Refine subsampling for late SC2 dataset
Removes prioritization of recombinant lineages now that we explicitly
include major recombinant lineages and their parental lineages in a
separate step. Increases the number of samples per lineage from that
known recombinant/parental set to 20 for a max of 20*25 or 500 total
samples and decreases the remaining subsampled set to 1500 samples for a
max of 2000 samples.