Filter cluster-specific mutations by allele frequency
Adds an argument to the cluster-specific mutation script and associated
Snakemake rules to filter mutations to a minimum allele frequency in
addition to the existing filter to a minimum allele count. This
additional filter now retains only mutations that occur in at least 50%
of strains in a given cluster, increasing the representativeness of
mutations assigned to clusters.