Revert Augur version update to avoid GFF error
Somewhere between Augur 22.0.3 and 22.4.0, augur export v2 stopped
supporting amino acid translations produced by augur translate from the
GFF file we use for the SC2 gene map in this analysis. This is the same
GFF file used by the ncov workflow. This commit reverts to the older
version of Augur that allowed the tree to be exported. However, I
noticed that this version still does not parse the GFF properly, only
keeping the last item in the file (the S gene) and missing the other
genes.