Add rules to embed representative simulated data
Adds rules to select the first four years of simulated data from a given
virus and replicate and create embeddings for those data with the
optimal parameters from the grid search. This is the first step toward
creating a plot of representative embeddings per virus and method with
the optimal parameters.