Update optimal cluster results for early flu
Re-running the early flu workflow can produce slightly different VI
values which suggests we haven't pinned random seeds everywhere in the
workflow or that those seeds aren't working as expected. Since VI values
have changed here for three of the methods including PCA which should be
deterministic, it's possible that the variation is in the input data
themselves.