Update TreeKnit steps of HA/NA analysis
Updates the HA/NA workflow to use the latest recommended steps for
TreeKnit analyses by rooting both HA and NA divergence trees on the same
strain, optionally collapsing nodes with low support (defaulting here to
support of 1), and running TreeKnit with the `--better-MCCs` flag and the
newer recommended value of gamma for H3N2. Also, updates the refine rule
to use FFT and stochastic polytomy resolution for faster and better
trees, respectively.