Fix color scale for MCCs
Use the default (essentially random) order for MCCs instead of sorting
them by name and assign colors to them in this order before sorting them
by name for display in the final figure legend. This approach keeps the
largest MCCs from having overly similar colors that make them difficult
to distinguish in the tree or embeddings.