Drop columns with N bases from HA/NA alignments
Adds a script and Snakemake rule to drop columns with N characters from
the HA and HA/NA alignments used by PCA (including when PCA is used to
initialize t-SNE embeddings) and updates rules in the HA/NA workflow to use
these cleaned alignments. This change does not affect distance-based
methods, since the distance matrix gets calculated from the original
alignments. However, the columns with Ns are uninformative for PCA
except to detect samples with Ns, so this change improves the biological
value of the PCA embeddings for HA and NA.