Omit internal nodes from cluster accuracy calc
Adds a rule to filter internal nodes from the table that we use as
input to cluster accuracy calculations. Including internal nodes falsely
inflates the number of "true clusters", but it has no effect on the VI
calculations. This change should only be needed in the early flu
workflow, since the early SC2 workflow uses predefined clade labels in
the metadata.