Add replication analysis for late SC2
Also, use only 5 replicates per dataset. Note that the SC2 analysis is
slowed down substantially by the time to run augur filter for each
subsample. We could speed this up by manually splitting the metadata by
region, year, and month and randomly sampling from these bins to get
strains in parallel. This would increase the complexity of the workflow,
though, so it may not be worth the effort.