Update data for SC2 analyses
Updates instructions for preparing SC2 data to reflect the need to save
both the filtered metadata and alignments in a S3 path specific to this
project. We need to save alignments, too, since the upstream open
Nextstrain alignments can change over time and have done so recently
such that our metadata in this project's S3 path didn't match the open
alignments available today. This commit also updates the date-stamp in
the URLs referenced in the SC2 workflows.