Add workflow to test effects of genetic bias
Adds workflow for late seasonal flu data to test the effects of genetic
bias on embedding methods. Runs the same general steps as the main late
flu workflow (alignment, tree, embeddings, clusters), calculates the
accuracy of the embeddings vs. Nextstrain clades for comparison with the
late flu data, and plots the relationship between Euclidean and genetic
distances in the biased sample.