Dynamically make list of flu clades to plot
Instead of hardcoding a list of flu clades to plot in early and late
data, build the list of clades dynamically from the input table. For the
later flu data, only assign colors to clades with at least 10 samples,
so we can more easily distinguish these larger groups and not waste the
few colors we have on smaller groups.