Calculate embedding positions for internal nodes
Adds a script and rule to the early flu workflow to calculate embedding
positions for internal nodes in the tree based on the mean position of
the immediate descendants of each internal node. The result of this
change is that internal nodes have positions in the Auspice JSONs and
embedding data CSV produced by the workflow. These changes allow users
to browse embeddings by branches in Auspice without any influence from
the phylogenetic inference on the embedding positions. We can also use
the positions in the plotting logic to plot branches in our static and
interactive figures.