WIP: Prototype plotting branches on to embeddings
Initial changes to the flu plotting logic to calculate positions for
internal nodes and use these positions to build line segments to plot as
branches for each embedding. This prototype needs to use the
precalculated internal node positions from the previous commit to build
line segments. We also need to extend this logic to the late flu and
early SC2 workflows.

Related to #78