Quantify monophyletic nature of clusters
Adds workflow rules and a script to summarize the monophyletic nature of
embedding clusters by first running DTA on cluster labels and then
counting the transitions between inferred cluster label states between
internal nodes of the tree. Reports the number of extra transitions that
appear in the tree for all transitions along with the total number of
clusters and transitions.

This commit adds rules to the early and late H3N2 HA workflows.
Subsequent commits will add similar rules to the H3N2 HA/NA and
early/late SC2 workflows.

Closes #79