Add random subsampling to late flu replication analysis
Adds a wildcard to the late flu workflow's replication analysis to
produce subsampling with random sequences in addition to even
subsampling. Updates the number of sequences sampled to a max total
number instead of sequences per group, to make the number of sequences
in each replicate comparable across subsampling schemes.