Clarify importance of linearity in embeddings
Adds references to studies that disagree about the meaning of global
structures in PCA, t-SNE, and UMAP and note that one goal of our project
is to find out how meaningful those embeddings are for viral genome
sequences.

Updates the title of the first results section to highlight the key
finding that method parameters have little effect on the global
structures in each embedding method and updates the first paragraph
of that section to clarify that we maximize the linear relationship
between Euclidean and genetic distances to maximize the interpretability
of the visualizations.

Closes #86