Use GitHub Actions to deploy repo website
Replaces the "classic" GitHub Pages deployment (which only allows us to deploy files from a `docs/` directory) with the newer GitHub Actions approach. This commit adds the simpler workflow to deploy static content instead of the more complex Jekyll-based workflow. This approach allows us to copy all of the content from the `manuscript/figures/` directory to the repo's website, so we can link to the interactive figures from the repo's README. We could also eventually add an index page in the figures directory that links to these figures, too.