Replace seaborn boxplot with matplotlib
Fixes a subtle bug in the boxplots of Euclidean distance by genetic
distance where a genetic distance of 76 was plotted directly next to a
distance of 70 because the boxplot x-axis is expected to be categorical.
This commit replaces the seaborn boxplot function with the matplotlib
function which allows us to specify a continuous range of genetic values
to plot on the x-axis and show gaps between values on that axis when not
all genetic distances are present.

A subsequent commit will add the updated boxplot figures for the
manuscript based on the change in this commit.

Related to #89