Try simplifying Docker image builds
Uses the approach from @victorlin's github-actions-docker-build [1]
template to create a Docker image for this project with GitHub Actions
and push the resulting image to the GHCR. This first attempt replaces
the more complicated previous Dockerfile with one that inherits from the
Nextstrain base image which has almost all of the dependencies we need
for this project installed now. Then, we install the few remaining
packages that aren't already in that base image including Python
packages and Java. Still to be added are Julia and TreeKnit.

[1] https://github.com/victorlin/github-actions-docker-build