Sort distance and embedding inputs before accuracy calculation
Fixes a bug in the grid search accuracy calculation introduced in pathogen-embed
2.0.0 where the embedding output were getting sorted by strain name and the
genetic distances that we compare to the embeddings were not getting sorted. The
proximal fix is implemented here where we sort all of the inputs by strain name
before calculating the accuracy of the train/test data. The ultimate fix is to
undo the internal sorting by strain name in pathogen-embed so genetic distances
and embeddings don't get unlinked.