Process HA and NA alignments separately
Replaces logic for running distance calculations and embeddings on
either an HA alignment or a concatenated HA/NA alignment with logic to
get distances and embeddings from HA and NA alignments separately. This
works because pathogen-embed supports multiple input values to its
alignment and distance matrix arguments. The benefit of this change is
that the pathogen-distance command can calculate distances that ignore
leading and trailing gaps in each gene's alignment that would otherwise
be counted in the concatenated alignment. Since we did not calculate
indel distances for the HA/NA analysis, this change to the workflow
should only affect the PCA embeddings. Since the new simplex encoding of
PCA inputs effectively ignores gaps, even the PCA embeddings should be
minimally affected by this change. However, the most important aspect of
this change is the demonstration of how we recommend these tools to be
used for this kind of reassortment analysis.

Closes #121