Plot trees with rectangular layout
Replaces diagonal line segments between parent and child nodes in the
tree with separate vertical and horizontal line segments to create a
more traditional phylogenetic tree representation. Updates the placement
of clade labels in the H3N2 tree plots based on this new layout and adds
an option to disable coloring branches when it doesn't make sense to
color them (e.g., when coloring tree by cluster labels for sequences).

Closes #106