Add logic to plot number of group-specific muts
Adds workflow logic and a script to plot the distribution of
total group-specific mutations per pathogen dataset and genetic group
using the supplemental table of mutations per cluster. Adds the
resulting figure as a supplemental figure in the manuscript and
references this figure wherever we already reference the corresponding
supplemental table.