Reposition overlapping clade labels
Adds logic to drop or reposition clade labels and then updates all
figures that had overlapping labels to prevent overlap. In extreme
cases when multiple clade labels overlap unavoidably in an embedding, we
pick the most ancestral clade's label to show and note this choice in
the caption of each affected figure.