Fix graphics include statements and include supplement in main
Use the preferred "!ht" instead of "!h" for placement (since LaTeX
automatically converts to this former string anyway) and define the
graphics path and includegraphics paths for the document to source
figures from either the same directory as the main TeX file or the
figures directory. This last change allowed the journal's submission
portal to properly render figures.

Also copies the supplemental data legends into the main document and
makes the supplement its own standalone document. This change allows the
references in the main doc to the supplement to work without including
the figures and tables and also allows us to build the supplement as an
independent PDF to upload.