Use "H" to place figs/tables and fix float error
Replaces the "[!ht]" placement codes for figures and tables with "[H]"
which causes these environments to be placed where they appear in the
text and not be treated as floats. This fixes a bug during submission to
the journal where floats would get "lost" for half of the supplementary
figures, breaking references to these figures in the main text.